+++ sdk/identity/identity/src/msal/nodeFlows/msalClient.ts
@@ -465,6 +465,17 @@ To work with multiple accounts for the same Client ID and Tenant ID, please prov
   }
 
   /**
+   * Builds an authority URL for the given request. The authority may be different than the one used when creating the MSAL client
+   * if the user is creating cross-tenant requests
+   */
+  function calculateRequestAuthority(options?: GetTokenOptions): string | undefined {
+    if (options?.tenantId) {
+      return getAuthority(options.tenantId, createMsalClientOptions.authorityHost);
+    }
+    return state.msalConfig.auth.authority;
+  }
+
+  /**
    * Performs silent authentication using MSAL to acquire an access token.
    * If silent authentication fails, falls back to interactive authentication.
    *
@@ -532,7 +543,7 @@ To work with multiple accounts for the same Client ID and Tenant ID, please prov
     try {
       const response = await msalApp.acquireTokenByClientCredential({
         scopes,
-        authority: state.msalConfig.auth.authority,
+        authority: calculateRequestAuthority(options),
         azureRegion: calculateRegionalAuthority(),
         claims: options?.claims,
       });
@@ -563,7 +574,7 @@ To work with multiple accounts for the same Client ID and Tenant ID, please prov
     try {
       const response = await msalApp.acquireTokenByClientCredential({
         scopes,
-        authority: state.msalConfig.auth.authority,
+        authority: calculateRequestAuthority(options),
         azureRegion: calculateRegionalAuthority(),
         claims: options?.claims,
         clientAssertion,
@@ -594,7 +605,7 @@ To work with multiple accounts for the same Client ID and Tenant ID, please prov
     try {
       const response = await msalApp.acquireTokenByClientCredential({
         scopes,
-        authority: state.msalConfig.auth.authority,
+        authority: calculateRequestAuthority(options),
         azureRegion: calculateRegionalAuthority(),
         claims: options?.claims,
       });
@@ -625,7 +636,7 @@ To work with multiple accounts for the same Client ID and Tenant ID, please prov
         scopes,
         cancel: options?.abortSignal?.aborted ?? false,
         deviceCodeCallback,
-        authority: state.msalConfig.auth.authority,
+        authority: calculateRequestAuthority(options),
         claims: options?.claims,
       };
       const deviceCodeRequest = msalApp.acquireTokenByDeviceCode(requestOptions);
@@ -654,7 +665,7 @@ To work with multiple accounts for the same Client ID and Tenant ID, please prov
         scopes,
         username,
         password,
-        authority: state.msalConfig.auth.authority,
+        authority: calculateRequestAuthority(options),
         claims: options?.claims,
       };
 
@@ -693,7 +704,7 @@ To work with multiple accounts for the same Client ID and Tenant ID, please prov
         scopes,
         redirectUri,
         code: authorizationCode,
-        authority: state.msalConfig.auth.authority,
+        authority: calculateRequestAuthority(options),
         claims: options?.claims,
       });
     });
@@ -725,7 +736,7 @@ To work with multiple accounts for the same Client ID and Tenant ID, please prov
     try {
       const response = await msalApp.acquireTokenOnBehalfOf({
         scopes,
-        authority: state.msalConfig.auth.authority,
+        authority: calculateRequestAuthority(options),
         claims: options.claims,
         oboAssertion: userAssertionToken,
       });
@@ -802,7 +813,7 @@ To work with multiple accounts for the same Client ID and Tenant ID, please prov
           await interactiveBrowserMockable.open(url, { wait: true, newInstance: true });
         },
         scopes,
-        authority: state.msalConfig.auth.authority,
+        authority: calculateRequestAuthority(options),
         claims: options?.claims,
         loginHint: options?.loginHint,
         errorTemplate: options?.browserCustomizationOptions?.errorMessage,